feat: moved dictionary and make some string changeable
