+++ patches/src/main/kotlin/app/revanced/patches/revenuecat/premium/UnlockProPatch.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.all.misc.revenuecat
+package app.revanced.patches.revenuecat.premium
 
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.patch.bytecodePatch
@@ -10,7 +10,29 @@ import com.android.tools.smali.dexlib2.iface.reference.StringReference
 @Suppress("unused")
 val unlockProPatch = bytecodePatch(
     name = "Unlock RevenueCat Pro",
-) {
+)
+
+{
+    val entitlementsName by stringOption(
+        key = "entitlements_name",
+        default = "Premium",
+        title = "Entitlement Name",
+        description = "The name of the entitlement to unlock. Default is 'Premium'. This could be found in v1/product_entitlement_mapping or /offerings",
+    )
+
+    val subscriptionsName by stringOption(
+        key = "subscriptions_name",
+        default = "Premium",
+        title = "Subscriptions Name",
+        description = "The name of the subscriptions to unlock. Default is 'Premium'. This could be found in v1/product_entitlement_mapping or /offerings",
+    )
+
+    val product_identifier by stringOption(
+        key = "product_identifier",
+        default = "premium",
+        title = "Product Identifier",
+        description = "The product identifier for the subscription. Default is 'premium'. This could be found in v1/product_entitlement_mapping or /offerings",
+    )
     execute {
         proFingerprint.method.apply {
             val subscriptionsInstruction = implementation!!.instructions.first { it ->
@@ -28,7 +50,7 @@ val unlockProPatch = bytecodePatch(
             addInstructions(
                 subscriptionsInstruction.location.index + 3,
                 """
-                    const-string v4, "{\"pro_access2\":{\"auto_resume_date\":null,\"billing_issues_detected_at\":null,\"expires_date\":\"2099-12-31T00:00:00Z\",\"grace_period_expires_date\":null,\"is_sandbox\":false,\"original_purchase_date\":\"2000-01-01T00:00:00Z\",\"period_type\":\"normal\",\"product_plan_identifier\":\"monthly-plan2\",\"purchase_date\":\"2000-01-01T00:00:00Z\",\"refunded_at\":null,\"store\":\"play_store\",\"unsubscribe_detected_at\":null}}"
+                    const-string v4, \"\${\"$product_identifier\":{\\\"auto_resume_date\\\":null,\\\"billing_issues_detected_at\\\":null,\\\"expires_date\\\":\\\"2099-12-31T00:00:00Z\\\",\\\"grace_period_expires_date\\\":null,\\\"is_sandbox\\\":false,\\\"original_purchase_date\\\":\\\"2000-01-01T00:00:00Z\\\",\\\"period_type\\\":\\\"normal\\\",\\\"product_plan_identifier\\\":\\\"monthly-plan2\\\",\\\"purchase_date\\\":\\\"2000-01-01T00:00:00Z\\\",\\\"refunded_at\\\":null,\\\"store\\\":\\\"play_store\\\",\\\"unsubscribe_detected_at\\\":null}}\"
                     new-instance v6, Lorg/json/JSONObject;
                     invoke-direct {v6, v4}, Lorg/json/JSONObject;-><init>(Ljava/lang/String;)V
                     const-string v4, "subscriptions"
@@ -38,7 +60,7 @@ val unlockProPatch = bytecodePatch(
             addInstructions(
                 entitlementsInstruction.location.index + 3,
                 """
-                    const-string v9, "{\"Pro\":{\"expires_date\":\"2099-12-31T00:00:00Z\",\"grace_period_expires_date\":null,\"product_identifier\":\"pro_access2\",\"product_plan_identifier\":\"monthly-plan2\",\"purchase_date\":\"2000-01-01T00:00:00Z\"}}"
+                    const-string v9, \"{\\\"$entitlementsName\\\":{\\\"expires_date\\\":\\\"2099-12-31T00:00:00Z\\\",\\\"grace_period_expires_date\\\":null,\\\"product_identifier\\\":\\\"$product_identifier\\\",\\\"product_plan_identifier\\\":\\\"monthly-plan2\\\",\\\"purchase_date\\\":\\\"2000-01-01T00:00:00Z\\\"}}\"
                     new-instance v8, Lorg/json/JSONObject;
                     invoke-direct {v8, v9}, Lorg/json/JSONObject;-><init>(Ljava/lang/String;)V
                 """
